+++ crawler/relevance.go
@@ -1,5 +1,5 @@
 package crawler
 
 type RelevanceFilter interface {
-	IsURLRelevant(content string) (bool, float64, error)
+	IsURLRelevant(content string) (bool, float32, error)
 }
